+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## [1.0.4]
+* Set the global currentChatUserId being chat with.
+* Fix IMMessage entry, add fromName field.
+* Improve documentation。
+
 ## [1.0.3]
-* Modify some class name. 
+* Modify some class name.
 
 ## [1.0.2]
 * export necessary bean.